getSkillsByTags deleted from controller, controller reorganized, mapper and dto added

+++ src/main/java/com/hse/skillsevaluation/service/SkillServiceImpl.java
@@ -1,11 +1,12 @@
 package com.hse.skillsevaluation.service;
 
 import com.hse.skillsevaluation.entity.skill.Skill;
+import com.hse.skillsevaluation.exception_handling.NoSuchSkillException;
 import com.hse.skillsevaluation.repositry.SkillRepository;
 import java.util.List;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -17,23 +18,37 @@ public class SkillServiceImpl implements SkillService {
     return skillRepository.findAll();
   }
 
+  @Transactional
+  @Override
+  public List<Skill> saveAllSkills(List<Skill> skills) {
+    return skillRepository.saveAll(skills);
+  }
+
   @Override
   public Skill getSkillById(Long id) {
-    Skill skill = null;
-    Optional<Skill> optionalEmployee = skillRepository.findById(id);
-    if (optionalEmployee.isPresent()) {
-      skill = optionalEmployee.get();
-    }
-    return skill;
+    return skillRepository.findById(id).orElseThrow(NoSuchSkillException::new);
   }
-  
+
   @Override
-  public void saveSkill(Skill skill) {
-    skillRepository.save(skill);
+  public Skill saveSkill(Skill skill) {
+    if (skillRepository.existsById(skill.getId())) {
+      return skillRepository.save(skill);
+    } else {
+      throw new NoSuchSkillException();
+    }
+  }
+
+  @Override
+  public Skill addSkill(Skill skill) {
+    return skillRepository.save(skill);
   }
 
   @Override
   public void deleteSkillById(Long id) {
-    skillRepository.deleteById(id);
+    if (skillRepository.existsById(id)) {
+      skillRepository.deleteById(id);
+    } else {
+      throw new NoSuchSkillException();
+    }
   }
 }
